Split Soc.scala
